refactor: Правки комментариев
