+++ src/app/app-navigation/app-navigation.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { styled } from '@shared/ui/theme'
-import { SnackConnector } from '@features/snack-connector'
 import { TabsNavigation } from 'processes/navigation'
 import axios from 'axios'
+import { SnackConnector } from '@entities/snack-connector'
 
 const Wrapper = styled.View`
   flex: 1;
