+++ src/entities/snack-connector/index.ts
