+++ src/entities/snack-connector/model/index.ts
@@ -1,0 +1,3 @@
+export { $snackStore, showSnack, hideAllSnack, hideTopSnack } from './model'
+export type { TSnack, SnackType } from './model'
+export { colorHelper } from './color-helper'
