+++ src/entities/snack-connector/model/model.ts
@@ -1,17 +1,19 @@
 import { createEvent, createStore, sample } from "effector";
 
-export type SnackModel = {
-    type: 'error' | 'successes' | 'warning'
+export type SnackType = 'error' | 'successes' | 'warning'
+
+export type TSnack = {
+    type: SnackType
     message: string
     duration: number
 }
 
-export const $snackStore = createStore<SnackModel[]>([])
+export const $snackStore = createStore<TSnack[]>([])
 
-export const showSnack = createEvent<SnackModel>()
+export const showSnack = createEvent<TSnack>()
 export const hideTopSnack = createEvent()
 export const hideAllSnack = createEvent()
 
 $snackStore.on(showSnack, (state, payload) => ([...state, payload]))
-$snackStore.on(hideTopSnack, (state, _) => state.filter((_, index) => index !== 0))
+$snackStore.on(hideTopSnack, (state, _) => state.slice(1))
 $snackStore.reset(hideAllSnack)
