+++ src/entities/snack-connector/ui/index.ts
