--- src/features/snack-connector/model/index.ts
@@ -1,1 +1,0 @@
-export { $snackStore, showSnack, hideAllSnack, hideTopSnack } from './model'
